==> Desktop/CASSIOPEIA/diss_koppold/script/03_BBprepro.Rmd <==
---
title: "03_BBprepro"
author: "Alina Koppold"
date: '2022-07-04'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(signal)
library(ggplot2)
library(dplyr)
```

```{r load-data}
load("./data/merged/behave_bb.RData")
```

```{r single-subject-bandpassfiltering}
sub45 = subset(behave_bb, ID == 45) %>% select(ID, COPY, trialnr)
class(sub45$COPY)
```

```{r filtering-seewave}
f = 100 # sampling rate
low = 0.01
high = 10
sub45$res = seewave::bwfilter(sub45$COPY, f = f, from = low, to = high, bandpass = T, n= 4)

plot(sub45$COPY, type = "l", col = "grey", main = "seewave_pass")
lines(sub45$res, col = "blue")
```

```{r filtering-signal}
bf = butter(n =4, W= c(low/50, high/50), type = c("pass"))
freqz(bf, Fs = 100)

sub45$res2 = signal::filter(bf, sub45$COPY)

plot(sub45$COPY, type = "l", col = "grey", main = "signal_butterworth")
lines(sub45$res2, col = "red")

```

```{r filtering-dplR}
sub45$res3 = dplR::pass.filt(sub45$COPY, W = c(low/50, high/50), type = "pass")
plot(sub45$COPY, type = "l", col = "grey", main = "dplR_bw")
lines(sub45$res3, col = "yellow")
```

```{r filtering-bb-data}
behave_bb$COPYbp = signal::filter(bf, behave_bb$COPY)
behave_bb$COPXbp = signal::filter(bf, behave_bb$COPX)

```

```{r n-observation-check}

# a. checking observations each subject
# mean observation (rows) +/- 2SD
df_obsCount= behave_bb %>% group_by(ID) %>% summarise(nObs = n())
df_obsCount$Mean =mean(df_obsCount$nObs, na.rm = T)
df_obsCount$std =sd(df_obsCount$nObs, na.rm = T)
df_obsCount$outliercheck = ifelse(df_obsCount$nObs > df_obsCount$Mean + (2*df_obsCount$std) | df_obsCount$nObs < df_obsCount$Mean - (2*df_obsCount$std), "outlier", "check")
table(df_obsCount$outliercheck)


```

# Dataquality
 compare every datapoint of a trial with mean datapoint value above trials within subject
 more than 50% datapoint outliers of a trial --> exclusion 
 how many trials can be excluded to still include subject?

```{r single-sub-quality}

s45 = subset(behave_bb, ID == 45)

# numbering datapoints
s45 = s45 %>% group_by(trialnr) %>% mutate(rown = row_number())

# mean and sd for datapoints over trials within subject
s45_info = s45 %>% group_by(rown) %>% summarise(MeanY = mean(COPYbp, na.rm = T), sdY = sd(COPYbp, na.rm = T), num = n())
s45 = merge(s45, s45_info, by = "rown")

# identification of trial-outliers within subject
s45$outliercheck = ifelse(s45$COPYbp > s45$MeanY + (2*s45$sdY)| s45$COPYbp < s45$MeanY - (2*s45$sdY), "outlier", "check")
a = table(s45$outliercheck)
prop.table(a)

s45$trialnr = as.factor(s45$trialnr)
everyother <- function(x) x[seq_along(x) %% 10 == 0]

# viz 
t1 = ggplot(data = s45, aes(x = trialnr, y = COPYbp, color= outliercheck))+
  geom_line()+
  geom_vline(xintercept = 30.5, color = "grey")+
  geom_vline(xintercept = 60.5, color = "grey")+
  scale_x_discrete(breaks = everyother)+
  scale_color_manual(values= c( "black", "red"))+
  geom_text(aes(x = 38, label = "BREAK 1", y = 0.1), color = "grey", angle = 0)+
  geom_text(aes(x = 68, label = "BREAK 2", y = 0.1), color = "grey", angle = 0)+
  theme_classic() + theme(axis.text.x = element_text(size = 14),
                          axis.text.y = element_text(size = 14))
t1
```

```{r datapoint-distribution within trial}
te = s45 %>% group_by(ID, trialnr, outliercheck) %>% summarise(freq = n())
te_info = te %>% group_by(trialnr) %>% summarise(Sum_datapointspertrial = sum(freq))
te = merge(te, te_info, by = "trialnr")
outlier_frame = subset(te, outliercheck == "outlier")
outlier_frame$remove = ifelse(0.5 <= outlier_frame$freq / outlier_frame$Sum_datapointspertrial, "delete", "check")
outlier_frame$outl_ratiowithintrial = outlier_frame$freq / outlier_frame$Sum_datapointspertrial

# more than 50% datapoint outliers of a trial --> exclusion
# checking how many trials are outlier
b = table(outlier_frame$remove)
prop.table(b)

```

```{r dataquality-all-subjects}

outlier_all = data.frame()

for (i in 0:length(unique(behave_bb$ID))){ #0
  # if(i == 8 || i == 19|| i == 20 || i == 37 || i == 58 || i == 62)
  #   next
  s5 = subset (behave_bb, ID == i)
  subject = s5 %>% select(ID) %>% unique()
  subject = as.vector(subject)
  subject = subject[[1]]
  print (subject)
  
  # numbering datapoints
  s5 = s5 %>% group_by(trialnr) %>% mutate (rown = row_number())
  
  # create mean and sd for datapoints within subject
  s5_info = s5 %>% group_by(rown) %>% summarise(MeanY = mean(COPYbp, na.rm = T), sdY = sd(COPYbp,
  na.rm = T), num = n())
  s5 = merge(s5, s5_info, by = "rown") 
  # identify outliers within subject per trial
  s5$outliercheck = ifelse(s5$COPYbp > s5$MeanY + (2*s5$sdY) | s5$COPYbp < s5$MeanY -s5$sdY,
  "outlier",  "check" )
  a =table (s5$outliercheck)
  prop.table(a)
  print(a)
  
  #visualise
  t2 = ggplot(data = s5, aes(x=trialnr, y= COPYbp, color=outliercheck))+
    geom_line() + theme_classic() +
    ggtitle("Subject", subject)+
    scale_color_manual(values= c( "black", "red"))+
    geom_line(aes(x = trialnr, y= MeanY), color="cyan1" ) +
    geom_vline(xintercept = 30.5, colour = "grey") + 
    geom_vline(xintercept = 60.5, colour = "grey")+
    geom_text(aes(x=38, label= "BREAK 1", y=0.1), colour="grey",angle=0) +
    scale_x_discrete(breaks = everyother)+
    geom_text(aes(x=68, label="BREAK 2", y=0.1), colour="grey", angle=0)
  print(t2)
  
  # identify outliers within trial
  te = s5 %>% group_by(ID, trialnr, outliercheck) %>% summarise(freq = n())
  te_info = te %>% group_by(trialnr) %>% summarise(Sum_datapointspertrial = sum(freq))
  te = merge(te, te_info, by = "trialnr")
  outlier_frame = subset(te, outliercheck == "outlier")
  outlier_frame$remove = ifelse(0.5 <= outlier_frame$freq / outlier_frame$Sum_datapointspertrial, "delete", "check")
  outlier_frame$outl_ratiowithintrial = outlier_frame$freq/ outlier_frame$Sum_datapointspertrial
  # more than 50% datapoint outliers of a trial -› trial exclusion
  # checking how many trials are outlier
  b = table(outlier_frame$remove)
  prop.table(b)
  print(b)
  outlier_all = rbind(outlier_all, outlier_frame)
}
```

```{r clean-COPY}
outlier_all = outlier_all %>% filter(remove == "delete")
df_BB_clean_Y = anti_join(behave_bb, outlier_all, by = c("ID", "trialnr"))
df_BB_clean_Y$COPX = NULL
df_BB_clean_Y$COPXbp = NULL
```


```{r dataquality-all-subjects}

outlier_allx = data.frame()

for (i in 0:length(unique(behave_bb$ID))){ #0
  # if(i == 8 || i == 19|| i == 20 || i == 37 || i == 58 || i == 62)
  #   next
  s5x = subset (behave_bb, ID == i)
  subject = s5x %>% select(ID) %>% unique()
  subject = as.vector(subject)
  subject = subject[[1]]
  print (subject)
  
  # numbering datapoints
  s5x = s5x %>% group_by(trialnr) %>% mutate (rown = row_number())
  
  # create mean and sd for datapoints within subject
  s5x_info = s5x %>% group_by(rown) %>% summarise(MeanX = mean(COPXbp, na.rm = T), sdX = sd(COPXbp, na.rm = T), num = n())
  s5x = merge(s5x, s5x_info, by = "rown") 
  # identify outliers within subject per trial
  s5x$outliercheck = ifelse(s5x$COPXbp > s5x$MeanX + (2*s5x$sdX) | s5x$COPXbp < s5x$MeanX -s5x$sdX,
  "outlier",  "check" )
  ax =table (s5x$outliercheck)
  prop.table(ax)
  print(ax)
  
  #visualise
  t2 = ggplot(data = s5x, aes(x=trialnr, y= COPXbp, color=outliercheck))+
    geom_line() + theme_classic() +
    ggtitle("Subject", subject)+
    scale_color_manual(values= c( "black", "red"))+
    geom_line(aes(x = trialnr, y= MeanX), color="cyan1" ) +
    geom_vline(xintercept = 30.5, colour = "grey") + 
    geom_vline(xintercept = 60.5, colour = "grey")+
    geom_text(aes(x=38, label= "BREAK 1", y=0.1), colour="grey",angle=0) +
    scale_x_discrete(breaks = everyother)+
    geom_text(aes(x=68, label="BREAK 2", y=0.1), colour="grey", angle=0)
  print(t2)
  
  # identify outliers within trial
  tex = s5x %>% group_by(ID, trialnr, outliercheck) %>% summarise(freq = n())
  tex_info = tex %>% group_by(trialnr) %>% summarise(Sum_datapointspertrial = sum(freq))
  tex = merge(tex, tex_info, by = "trialnr")
  outlier_framex = subset(tex, outliercheck == "outlier")
  outlier_framex$remove = ifelse(0.5 <= outlier_framex$freq / outlier_framex$Sum_datapointspertrial, "delete", "check")
  outlier_framex$outl_ratiowithintrial = outlier_framex$freq/ outlier_framex$Sum_datapointspertrial
  # more than 50% datapoint outliers of a trial -› trial exclusion
  # checking how many trials are outlier
  bx = table(outlier_framex$remove)
  prop.table(bx)
  print(bx)
  outlier_allx = rbind(outlier_allx, outlier_framex)
}
```

```{r clean-COPX}
outlier_allx = outlier_allx %>% filter(remove == "delete")
df_BB_clean_X = anti_join(behave_bb, outlier_allx, by = c("ID", "trialnr"))
df_BB_clean_X$COPY = NULL
df_BB_clean_X$COPYbp = NULL

df_BB_clean = merge(df_BB_clean_X, df_BB_clean_Y)
save(df_BB_clean, file = "./data/merged/df_BB_clean.RData")

```
