==> Desktop/CASSIOPEIA/diss_koppold/script/04_bb_analyses.Rmd <==
---
title: "04_bb_analyses"
author: "Alina Koppold"
date: '2022-07-04'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(dplyr)
library(ggplot2)
```

```{r data-prep}
load("./data/merged/df_BB_clean.RData")
df_BB_clean = df_BB_clean %>% filter(blinktime == 90)
df_sd = df_BB_clean %>% group_by(ID, trialnr, blink, opponent) %>%  summarise(sd_x = sd(COPXbp), sd_y = sd(COPYbp))

```

```{r function}
data_summary <- function(data, varname, groupnames){
  require(plyr)
  summary_func <- function(x, col){
    c(mean = mean(x[[col]], na.rm=TRUE),
      sd = sd(x[[col]], na.rm=TRUE))
  }
  data_sum<-ddply(data, groupnames, .fun=summary_func,
                  varname)
  data_sum <- rename(data_sum, c("mean" = varname))
 return(data_sum)
}

alinatheme <- readRDS('./functions/alinatheme.rds')
```


Threat Imminence 

```{r}
target = c("anticipation", "guy", "opp")
df_ti =df_BB_clean %>% filter(marker %in% target) 

df_ti$current_time = as.POSIXct(df_ti$current_time, format = "%d/%m/%Y:%H:%M:%S")

df_ti$current_time <- format(df_ti$current_time, format = "%S")

df_ti = df_ti %>%
  group_by(ID, trialnr) %>%
  mutate(seconds = match(paste(current_time), unique(paste(current_time)))-1)
table(df_ti$seconds)
df_ti = df_ti %>% filter(seconds < 12)



# dplyr tricks testing 
df_test = df_ti %>% rowwise() %>%
 # arrange(ID, desc(trialnr)) %>%
 # group_by(ID, trialnr, seconds) %>%
  mutate(min_xy = min(COPXbp, COPYbp)) %>%
  select(min_xy, everything()) %>% #at beginning of frame
  as.data.frame() %>%
  mutate_if(is.character, toupper) %>%
  head()

df_sanity = df_test %>% count(seconds, name = "Seconds.Count") #add_count(seconds)
colorDF::colorDF(head(df_sanity), theme = "wb")


# continuing analysis
# implement second 0 (start), plot with 0 


df_ti$session= with(df_ti, ifelse(trialnr < 30, "session1", 
          ifelse(trialnr >= 30 & trialnr < 61, "session2", "session3"))) #%>% mutate(case_when)

#detach("package:plyr", unload = T)
df_thi = df_ti %>% group_by(ID, trialnr, blink) %>%  summarise(sd_x = sd(COPXbp), 
                                                               sd_y = sd(COPYbp),
                                                               seconds = unique(seconds),
                                                               opponent = unique(opponent))



# choose session: "s1", "s2", "s3", "all"
session_of_int = "all"
varofint = c("sd_x", "sd_y")
for (i in 1:length(varofint)) {
  if (session_of_int == "s2") {
    df_sub = df_thi %>% filter(between(trialnr, 31, 59))
  } else if (session_of_int == "s1") {
    df_sub = df_thi %>% filter(trialnr < 30)
  } else if (session_of_int == "s3") {
    df_sub = df_thi %>% filter(trialnr > 60)
  } else if (session_of_int == "all") {
    df_sub = df_thi
  } else {
    df_sub = 0
  }
  library(plyr)
  d5 = data_summary(df_sub,
                    varname = varofint[i],
                    groupnames = c("seconds", "blink"))
  
  p0 = ggplot(d5, aes(x = seconds, y = get(varofint[i]), group = blink)) +
    geom_line(aes(color = blink)) +
    geom_point(aes(color = blink)) +
    labs(title = varofint[i], x = "seconds within trial") +
    scale_color_manual(
      "Anticipation condition",
      values = c("black", "blue", "yellow"),
      labels = c("threat imminence", "uncertain", "safe")
    ) +
    alinatheme
  print(p0)
}

df_ti = df_ti %>% filter(ID == "81")

# choose session: "s1", "s2", "s3", "all"
session_of_int = "all"
varofint = c("COPXbp", "COPYbp")
for (i in 1:length(varofint)) {
  if (session_of_int == "s2") {
    df_sub = df_ti %>% filter(between(trialnr, 31, 59))
  } else if (session_of_int == "s1") {
    df_sub = df_ti %>% filter(trialnr < 30)
  } else if (session_of_int == "s3") {
    df_sub = df_ti %>% filter(trialnr > 60)
  } else if (session_of_int == "all") {
    df_sub = df_ti
  } else {
    df_sub = 0
  }
  library(plyr)
  p1 = ggplot(df_sub, aes(x = seconds, y = get(varofint[i]), color = blink)) +
    geom_smooth()+
    labs(title = varofint[i], x = "seconds within trial") +
    scale_color_manual(
      "Anticipation condition",
      values = c("black", "blue", "yellow"),
      labels = c("threat imminence", "uncertain", "safe")
    ) 
  print(p1)
}
```



Habituation / initial value dependency

```{r bb-habituation}
# extract first trials
df_BB_clean$trialnr =as.numeric(as.character(df_BB_clean$trialnr))
detach("package:plyr", unload = TRUE)
getN = df_BB_clean %>% select(ID, trialnr) %>% group_by(ID, trialnr) %>% unique() %>% add_count()
df_ivd1 = df_BB_clean %>% filter(trialnr <=2)
#df_ivd1$ID = as.numeric(as.character(df_ivd1$ID))


df_ivd = df_ivd1   %>% group_by(ID, trialnr, blink) %>% summarise(sd_x = sd(COPXbp),sd_y = sd(COPYbp)) %>% arrange(ID, trialnr)

# spaghetti plot from ggvis package


varofint = list(c("sd_x", "change of lateral movement"), c("sd_y", "change of anterior movement"))


for (i in 1:length(varofint)) {
  #df_ivd$ID = as.factor(df_ivd$ID)
  s1 <-
    ggplot(data = df_ivd, aes(
      x = as.factor(trialnr),
      y = get(varofint[[i]][1]),
      group = ID,
      color = ID
    )) +
    alinatheme +
    labs(x = "trial", y = varofint[[i]][2]) +
    geom_line(size = 1) + theme(legend.position = "none") 
print(s1)
  
}
```

```{r baseline correction}
# subtrackting baseline 

```

```{r}
# plot difference in second (min max) as bin within trial / over trials 
#  slice_max(mass, n = 3, with_ties = F) %>%  #max difference of movement within second

```
